+++ app/Enums/BookingStatus.php
@@ -4,6 +4,11 @@ namespace App\Enums;
 
 use BenSampo\Enum\Enum;
 
+/**
+ * @method static static UNASSIGNED()
+ * @method static static RESERVED()
+ * @method static static BOOKED()
+ */
 final class BookingStatus extends Enum
 {
     public const UNASSIGNED = 0;