+++ app/Http/Requests/Airport/Admin/StoreAirport.php
@@ -17,6 +17,8 @@ class StoreAirport extends Request
             'icao' => 'required:string|unique:airports|size:4',
             'iata' => 'required:string|unique:airports,iata|size:3',
             'name' => 'required:string',
+            'latitude' => ['required', 'regex:/^[-]?((([0-8]?[0-9])(\.(\d{1,10}))?)|(90(\.0+)?))$/'],
+            'longitude' => ['required', 'regex:/^[-]?((((1[0-7][0-9])|([0-9]?[0-9]))(\.(\d{1,10}))?)|180(\.0+)?)/'],
         ];
     }
 
@@ -31,6 +33,8 @@ class StoreAirport extends Request
             'icao' => __('ICAO'),
             'iata' => __('IATA'),
             'name' => __('Name'),
+            'latitude' => __('Latitude'),
+            'longitude' => __('Longitude'),
         ];
     }
 }