+++ config/database.php
@@ -74,7 +74,7 @@ return [
             'charset' => 'utf8',
             'prefix' => env('DB_TABLE_PREFIX', ''),
             'prefix_indexes' => true,
-            'schema' => 'public',
+            'search_path' => 'public',
             'sslmode' => 'prefer',
         ],