+++ resources/js/alpine.js
@@ -1,0 +1,9 @@
+import Alpine from 'alpinejs';
+
+window.Alpine = Alpine;
+
+Alpine.start();
+
+
+import flatpickr from 'flatpickr';
+window.flatPickr = flatpickr;